Revert "[M3.c] normalize dsv4 swa mapping dtype once"
This reverts commit 99a246226de3c58a67aa0f545c61947e9114b494.

+++ python/sglang/srt/mem_cache/deepseek_v4_memory_pool.py
@@ -634,12 +634,7 @@ class DeepSeekV4TokenToKVPool(BaseSWAKVPool):
         self._should_cache_swa = envs.SGLANG_OPT_CACHE_SWA_TRANSLATION.get()
 
     def register_mapping(self, full_to_swa_index_mapping: torch.Tensor):
-        # The mapping is immutable after pool construction and all translated
-        # locations are consumed as int32. Normalize it once instead of
-        # launching a dtype conversion for every layer and decode step.
-        self.full_to_swa_index_mapping = full_to_swa_index_mapping.to(
-            dtype=torch.int32
-        )
+        self.full_to_swa_index_mapping = full_to_swa_index_mapping
 
     def get_ring_size(self, compress_ratio: int) -> int:
         server_args = get_global_server_args()
@@ -649,7 +644,7 @@ class DeepSeekV4TokenToKVPool(BaseSWAKVPool):
     def translate_loc_from_full_to_swa(self, kv_indices: torch.Tensor):
         assert self.full_to_swa_index_mapping is not None
 
-        return self.full_to_swa_index_mapping[kv_indices]
+        return self.full_to_swa_index_mapping[kv_indices].to(torch.int32)
 
     def set_swa_loc(self, loc: torch.Tensor) -> None:
         # No-op: SWAKVPool's set_swa_loc precomputes SWA-translated loc once per
